fix: move helper functions to the helper file
